requirements.txt
numpy
pandas
scikit-learn
pypyodbc
python-dotenv
#-e .